requirements.txt
asgiref==3.8.1
coverage==7.6.4
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
mysqlclient==2.2.5
sqlparse==0.5.1
